requirements.txt
cement==3.0.0
jinja2
pyyaml
colorlog
yaspin
colorama
emoji
pymongo
python-wifi
Pillow